pyproject.toml
[tool.poetry]
name = "interview-assessment"
version = "1.0.0"
description = "Interview Assessment"
authors = ["Engineering Team <engineering.team@progyny.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
PyMySQL = "^1.0.2"
cryptography = "^3.4.7"
sqlalchemy = "1.4.22"
pydantic = "^1.10.2"
numpy = "^1.23.4"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
